Refactor ModelInfo generation logic
